# Conflicts:
#   backend/archive/src/main/resources/application.yml

Former-commit-id: 9db428b70e0ab4ee0eac49515bd21ff39b1ad91a

+++ backend/archive/src/main/java/kr/co/reverse/archive/api/service/FriendService.java
@@ -60,8 +60,14 @@ public class FriendService {
             friendRepository.save(new Friend(target, user));
         }
 
-        FriendInvitation friendInvitation = friendInvitationRepository.findFriendInvitationByInvitationUserAndInvitationTarget(user, target);
-        friendInvitationRepository.delete(friendInvitation);
+        FriendInvitation friendInvitation = friendInvitationRepository.findFriendInvitationByInvitationUserAndInvitationTarget(target, user);
+        if(friendInvitation != null){
+            friendInvitationRepository.delete(friendInvitation);
+        }
+        FriendInvitation friendInvitation1 = friendInvitationRepository.findFriendInvitationByInvitationUserAndInvitationTarget(user, target);
+        if(friendInvitation1 != null){
+            friendInvitationRepository.delete(friendInvitation1);
+        }
     }
 
     @Transactional
